projectId, taskId 의 자료형을 Long -> Int로 수정, taskID를 project 소속과 관련 없이 고유하게 수정

+++ app/src/main/java/com/example/facemaker/MainActivity.kt
@@ -60,7 +60,7 @@ class MainActivity : AppCompatActivity() {
             }
         } else if (requestCode == projectDetailRequestCode && resultCode == Activity.RESULT_OK) {
             data?.let { data ->
-                val id = data.getLongExtra(REMOVED_PROJECT_ID, 0)
+                val id = data.getIntExtra(REMOVED_PROJECT_ID, 0)
                 ProjectManager.removeProjectForId(id)
             }
         }
